+++ WebNhutLong/WebNhutLong/Controllers/ProductsController.cs
@@ -2,10 +2,12 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
+using System.Globalization;
 using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
+using WebNhutLong.Domain;
 using WebNhutLong.Models;
 
 namespace WebNhutLong.Controllers
@@ -17,12 +19,20 @@ namespace WebNhutLong.Controllers
         // GET: Products
         public ActionResult Index()
         {
+            if (Session["username"]==null)
+            {
+                return RedirectToAction("Login", "Login");
+            }
             return View(db.tbl_Products.ToList());
         }
 
         // GET: Products/Details/5
         public ActionResult Details(int? id)
         {
+            if (Session["username"] == null)
+            {
+                return RedirectToAction("Login", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -34,12 +44,16 @@ namespace WebNhutLong.Controllers
             }
             return View(tbl_Products);
         }
-
+      
         // GET: Products/Create
         public ActionResult Create()
         {
+            
+        
             return View();
         }
+
+       
 
         // POST: Products/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
@@ -48,6 +62,10 @@ namespace WebNhutLong.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ID_Products,NameProducts,SolopProducts,QuyCachProducts,LoaigiayProducts,OffsetFlexoProducts,DanKimProducts,SoluongProducts,GiaProducts,CreatedDateProducts,ModifyDateProducts,CreateUserProducts,ModifyUserProducts,StatusProducts,CodeProducts")] tbl_Products tbl_Products)
         {
+            if (Session["username"] == null)
+            {
+                return RedirectToAction("Login", "Login");
+            }
             if (ModelState.IsValid)
             {
                 db.tbl_Products.Add(tbl_Products);
@@ -61,6 +79,10 @@ namespace WebNhutLong.Controllers
         // GET: Products/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session["username"] == null)
+            {
+                return RedirectToAction("Login", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -80,6 +102,10 @@ namespace WebNhutLong.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ID_Products,NameProducts,SolopProducts,QuyCachProducts,LoaigiayProducts,OffsetFlexoProducts,DanKimProducts,SoluongProducts,GiaProducts,CreatedDateProducts,ModifyDateProducts,CreateUserProducts,ModifyUserProducts,StatusProducts,CodeProducts")] tbl_Products tbl_Products)
         {
+            if (Session["username"] == null)
+            {
+                return RedirectToAction("Login", "Login");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(tbl_Products).State = EntityState.Modified;
@@ -92,6 +118,10 @@ namespace WebNhutLong.Controllers
         // GET: Products/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (Session["username"] == null)
+            {
+                return RedirectToAction("Login", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -109,12 +139,16 @@ namespace WebNhutLong.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Session["username"] == null)
+            {
+                return RedirectToAction("Login", "Login");
+            }
             tbl_Products tbl_Products = db.tbl_Products.Find(id);
             db.tbl_Products.Remove(tbl_Products);
             db.SaveChanges();
             return RedirectToAction("Index");
         }
-
+        
         protected override void Dispose(bool disposing)
         {
             if (disposing)
